Update package dependencies and Vidi software
Changed the 'extraneous' package-lock.json dependencies to development dependencies. Updated the starting quote and the copyright year in the index.html.default file. Also, updated the CHANGELOG.md with new recent changes, including addition of a new startup modal and comprehensive updates to the GUI in the new Vidi software release.