Add light theme option and import theme styles
Added a 'light' theme option to the initial configuration and loaded theme stylesheets in the main.scss file. Specifically, a 'theme' attribute was added to the 'window.vidiConfig' object and a 'data-bs-theme' attribute was set on the document root. A new 'themes' partial was also imported into the main Sass file.