Update URL conditional in init.js
Adjusted the URL conditional in the init.js file to include a check for urlVars.dfi (don't use fast init). This additional check ensures that the appropriate action takes place when the initial filter in the URL variables is not set and the dfi (Default Filter Index) is also not set.