Refactor and update snapshot-related components
The code has been refactored and updated to improve the handling of state snapshots. The `SearchFieldComponent.js` has been deleted, and `TitleFieldComponent.js` was renamed and moved to a new location. The functionality was also enhanced in `stateSnapshotsDashboard.js` and `stateSnapshots.js` which now support tags for state snapshots, introduced an enhanced filter for state snapshots, refactored the way it handles refreshing snapshots list, and improved the way it updates a snapshot.