Add support for 'data-vidi-schemata' attribute in embedded maps
The update now allows for an optional 'data-vidi-schemata' attribute in embedded maps. When this attribute is set, it specifies a valid config.js schemata text string. This addition enhances the customization of embedded maps, enabling the auto retrieval of matching layers alongside any specified in the config. Changes have been made in embed.js, meta.js and init.js to implement this new feature. Updates in the documentation were also made to reflect this addition.