2024.5.0 merge (#39)
* Set default value for getUri function parameter

The parameter 's' in the getUri function within anchor.js file has been updated to have a default value of null. This adjustment ensures the function will still work properly even if no argument is provided when calling this function.

* Refactor and update snapshot-related components

The code has been refactored and updated to improve the handling of state snapshots. The `SearchFieldComponent.js` has been deleted, and `TitleFieldComponent.js` was renamed and moved to a new location. The functionality was also enhanced in `stateSnapshotsDashboard.js` and `stateSnapshots.js` which now support tags for state snapshots, introduced an enhanced filter for state snapshots, refactored the way it handles refreshing snapshots list, and improved the way it updates a snapshot.

* Add 'db' prop to LayerFilterSubBlock

The 'db' prop has been added to the LayerFilterSubBlock component. This new prop is required, aligning to the convention used for other props in the same component. This ensures that the database is always defined for layer filters.

* Add tag filtering and creation in snapshot dashboard

Updated snapshot dashboard now supports tag filtering and creation functionality. Changes also include adding new translations and updating package-lock.json for new dependency requirements. Fixed duplication issues in the snapshot update process and updated snapshot and tag related styles for better UI experience.

* Refactor code and enable conflictSearch extension

The code has been refined to better handle null or undefined values when fetching the "gc2.host" value from vidiConfig. ConflictSearch extension is now enabled in the vidi.config.js file. The configSwitcher extension and some unrelated configurations are removed from the config file.

* New API for getting configs

* Update Dockerfile healthcheck retries

The healthcheck configuration in the Dockerfile was updated to include retries. A new argument is added to the healthcheck command to ensure a test will rerun twice if it fails the first time.

* The code change modifies `bindEvent.js` to delay the initialization of `switchLayer`. This change was accomplished by wrapping the `switchLayer.init` function inside a `setTimeout` function with a delay of 1ms. This may facilitate smoother loading and prevent potential blocking issues.

* Add layer visibility condition based on filter requirement

The code was updated to include a condition that checks if a filter is required for showing data on a layer. If a filter is necessary and has not been set, the layer will not be made visible. This feature helps to manage layer display more effectively, enhancing the user experience by only showing relevant content.

* Refactor snapshot filtering and improve code formatting

The snapshot filtering code has been refactored to a separate 'filter' function to enhance readability and consistency across different areas of the codebase. Additionally, code formatting has been improved in a few places for better visual clarity.

* Add "All on" and "All off" buttons to tags section

More functionalities were added to the StateSnapshotsDashboard.js file to improve the user interaction with the tags section. Users can now select or deselect all tags with the "All on" and "All off" buttons respectively.

* Update target container in bindEvent.js

The container ID for module events in bindEvent.js has been updated. The code now targets '#mainLayerOffcanvas' instead of '#module-container' to properly reflect changes in the HTML structure.

* 2024.5.0

* Fixed: Added deps back into build

* Update MarkupGenerator.js

Filter display property altered

* bug: matching layer _key_ when using jquery ^ operator

* Fixing matching layerKey  take

---------

Co-authored-by: Martin Høgh <mh@mapcentia.com>
Co-authored-by: Gunnar <gjj@geopartner.dk>
Co-authored-by: GunnarJulJensen <132446178+GunnarJulJensen@users.noreply.github.com>