Layout: Dropdown orientation
Navbar dropdown now defult to align-left, unless when session is enabled, then it goes back to align-right. This solves an issue where the user is unable to read the dropdown labels when session is not enabled in a config - prompting a "I can't read the menu because it is offscreen"-error.