Update offcanvas display condition and documentation
The condition for displaying the main layer offcanvas has been updated to account for mobile and screen width. The configuration documentation in the "run_configuration" file has also been updated accordingly to properly inform users about the new `showOffcanvas` and `expandFirstInLayerTree` attributes and their usage, and the minor correction was done for the "henvisning" description.